chore: query only the required commits (#72)
The rev parameter `<parent-hash>..<release-hash>` requests all the commits up to `<release-hash`> but excluding all the commits up to `<parent-hash>`. It queries all relevant commits regardless of `<parent-hash>` and `<release-hash>` being in the same branch.

It limits the amount of commits to save in memory and avoid using commits newer than the tag hash.