Support maintenance branch and hotfix release (#71)
* check tag instead of release

* Draft release points to the tag hash

* fix: support maintenance branch and hotfix release

Get the latest tag from the current branch. Allow to checkout a tag to release it or to checkout any branch to release the latest tag from that branch

* Get previous tag from all branches

Still query the release tag from the current branch, but query all tags to find the version preceding it; it might not be on the same branch.